cargando controllers y services

+++ BackEnd/elBuenSabor/src/main/java/com/peso/elBuenSabor/entities/ArticuloManufacturado.java
@@ -30,7 +30,7 @@ public class ArticuloManufacturado extends Base{
 
     @NotNull
     @Column(name = "precio_venta", precision = 10, scale = 2)
-    private BigDecimal precioVenta;
+    private Double precioVenta;
 
     @Column(name = "costo", precision = 10, scale = 2)
     private BigDecimal costo;
